indieauth/authorization_get_function/requirements.txt
chevron==0.13.1
PyJWT
cryptography
